pyproject.toml
[tool.poetry]
name = "projectmind"
version = "0.1.0"
description = "Autonomous app generator using LLM agents, LangGraph, and Supabase."
authors = ["Oscar Morera <olmorera@gmail.com>"]
readme = "README.md"
packages = [{ include = "projectmind" }]
package-mode = true

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
sqlalchemy = ">=2.0.41,<3.0.0"
alembic = ">=1.15.2,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
pydantic = ">=2.11.4,<3.0.0"
langgraph = "^0.4.5"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
python-dotenv = ">=1.1.0,<2.0.0"
slack-bolt = ">=1.23.0,<2.0.0"
loguru = ">=0.7.3,<0.8.0"
langgraph-checkpoint-postgres = "^2.0.21"
psycopg = {extras = ["binary"], version = "^3.2.9"}
langchain-core = "^0.3.60"
llama-cpp-python = "^0.3.9"
psutil = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
